rust: Split EKO to separate module
